refactor ustx plugin; ust reading (writing isn't currently implemented)
