+++ csharp/Console/OpenSvip.Console.csproj
@@ -22,8 +22,8 @@
     </PropertyGroup>
 
     <ItemGroup>
-      <PackageReference Include="CommandLineParser" Version="2.9.0-preview1" />
-      <PackageReference Include="NullLib.ConsoleTable" Version="1.0.3.2" />
+      <PackageReference Include="CommandLineParser"/>
+      <PackageReference Include="NullLib.ConsoleTable"/>
     </ItemGroup>
 
     <ItemGroup>
