+++ csharp/Directory.Packages.props
@@ -11,5 +11,16 @@
     <PackageVersion Include="Melanchall.DryWetMidi" Version="6.1.3" />
     <PackageVersion Include="System.Numerics.Vectors" Version="4.6.1" />
     <PackageVersion Include="System.Memory" Version="4.6.3" />
+    <PackageVersion Include="CommandLineParser" Version="2.9.1" />
+    <PackageVersion Include="NullLib.ConsoleTable" Version="1.0.3.2" />
+    <PackageVersion Include="MaterialDesignThemes" Version="4.5.0" />
+    <PackageVersion Include="Microsoft.Toolkit.Uwp.Notifications" Version="7.1.3" />
+    <PackageVersion Include="Microsoft.Xaml.Behaviors.Wpf" Version="1.1.39" />
+    <PackageVersion Include="Panuon.UI.Silver" Version="2.2.20" />
+    <PackageVersion Include="Samboy063.Tomlet" Version="3.1.3" />
+    <PackageVersion Include="WindowsAPICodePack-Core" Version="1.1.2" />
+    <PackageVersion Include="WindowsAPICodePack-Shell" Version="1.1.1" />
+    <PackageVersion Include="SharpZipLib" Version="1.4.2" />
+    <PackageVersion Include="NPinyin" Version="0.2.6321.26573" />
   </ItemGroup>
 </Project>
