+++ csharp/GUI/OpenSvip.GUI.csproj
@@ -164,27 +164,13 @@
     </ProjectReference>
   </ItemGroup>
   <ItemGroup>
-    <PackageReference Include="MaterialDesignThemes">
-      <Version>4.5.0</Version>
-    </PackageReference>
-    <PackageReference Include="Microsoft.Toolkit.Uwp.Notifications">
-      <Version>7.1.3</Version>
-    </PackageReference>
-    <PackageReference Include="Microsoft.Xaml.Behaviors.Wpf">
-      <Version>1.1.39</Version>
-    </PackageReference>
-    <PackageReference Include="Panuon.UI.Silver">
-      <Version>2.2.20</Version>
-    </PackageReference>
-    <PackageReference Include="Samboy063.Tomlet">
-      <Version>3.1.3</Version>
-    </PackageReference>
-    <PackageReference Include="WindowsAPICodePack-Core">
-      <Version>1.1.2</Version>
-    </PackageReference>
-    <PackageReference Include="WindowsAPICodePack-Shell">
-      <Version>1.1.1</Version>
-    </PackageReference>
+    <PackageReference Include="MaterialDesignThemes" />
+    <PackageReference Include="Microsoft.Toolkit.Uwp.Notifications" />
+    <PackageReference Include="Microsoft.Xaml.Behaviors.Wpf" />
+    <PackageReference Include="Panuon.UI.Silver" />
+    <PackageReference Include="Samboy063.Tomlet" />
+    <PackageReference Include="WindowsAPICodePack-Core" />
+    <PackageReference Include="WindowsAPICodePack-Shell" />
   </ItemGroup>
   <Import Project="$(MSBuildToolsPath)\Microsoft.CSharp.targets" />
 </Project>
